corosync: reload config rather than restarting
ensure_started is called once on xapi startup, and once per slave on xapi master
startup. When rebooting a pool, the restarting of corosync causes several
transients to occur before a settled cluster is formed. It may even cause hosts
to fence.

Instead, reload the corosync configuration without restarting the service. This
is much less invasive.

In case the host has changed pool, we also instruct corosync of the cluster
name, which may be different from what is saved in corosync.conf.

(This logically reverts the behaviour introduced in commit 6c1a1c77a.)

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>